Refactor project structure
Move some classes to other packages to improve management
