+++ lib/rules.pro
@@ -65,7 +65,7 @@
     @retrofit2.http.* <methods>;
 }
 
--keepclasseswithmembers class com.barista_v.debug_artist.report_bug.pivotal.Story { *; }
--keepclasseswithmembers class com.barista_v.debug_artist.report_bug.pivotal.StoryRequestBody { *; }
--keepclasseswithmembers class com.barista_v.debug_artist.report_bug.pivotal.Comment { *; }
--keepclasseswithmembers class com.barista_v.debug_artist.report_bug.pivotal.Attachment { *; }
+-keepclasseswithmembers class com.barista_v.debug_artist.repositories.pivotal.Story { *; }
+-keepclasseswithmembers class com.barista_v.debug_artist.repositories.pivotal.StoryRequestBody { *; }
+-keepclasseswithmembers class com.barista_v.debug_artist.repositories.pivotal.Comment { *; }
+-keepclasseswithmembers class com.barista_v.debug_artist.repositories.pivotal.Attachment { *; }
