+++ lib/src/main/java/com/barista_v/debug_artist/repositories/BugRepository.kt
@@ -4,11 +4,13 @@ import android.os.Parcelable
 import rx.Observable
 
 interface BugRepository {
-
-  fun createBug(name: String,
-                description: String,
-                screenshotFilePath: String?,
-                logsFilePath: String?): Observable<Answer<Any>>
+  /**
+   * Creates a bug
+   */
+  fun create(name: String,
+             description: String,
+             screenshotFilePath: String?,
+             logsFilePath: String?): Observable<Answer<Any>>
 
   interface Builder : Parcelable {
     fun build(): BugRepository
