+++ lib/src/main/java/com/barista_v/debug_artist/repositories/pivotal/PivotalBugRepositoryBuilder.kt
@@ -1,4 +1,4 @@
-package com.barista_v.debug_artist.report_bug.pivotal
+package com.barista_v.debug_artist.repositories.pivotal
 
 import android.os.Parcel
 import android.os.Parcelable
