+++ lib/src/main/java/com/barista_v/debug_artist/utils/device/AndroidDevice.kt
@@ -1,4 +1,4 @@
-package com.barista_v.debug_artist.utils
+package com.barista_v.debug_artist.utils.device
 
 import android.os.Process
 import android.support.v7.app.AppCompatActivity
