+++ lib/src/main/java/com/barista_v/debug_artist/utils/device/Device.kt
@@ -1,4 +1,4 @@
-package com.barista_v.debug_artist.utils
+package com.barista_v.debug_artist.utils.device
 
 interface Device {
   fun takeScreenshot(fileName: String): String
