+++ lib/src/main/java/com/barista_v/debug_artist/utils/shake_detector/AndroidShakeDetector.kt
@@ -1,4 +1,4 @@
-package com.barista_v.debug_artist.drawer.item.issue_reporter
+package com.barista_v.debug_artist.utils.shake_detector
 
 import android.content.Context
 import android.hardware.Sensor
